+++ b2sdk/bucket.py
@@ -13,7 +13,7 @@ import logging
 import pathlib
 
 from contextlib import suppress
-from typing import Optional, Tuple
+from typing import Dict, Optional, Tuple
 
 from .encryption.setting import EncryptionSetting, EncryptionSettingFactory
 from .encryption.types import EncryptionMode
@@ -37,11 +37,13 @@ from .progress import AbstractProgressListener, DoNothingProgressListener
 from .replication.setting import ReplicationConfiguration, ReplicationConfigurationFactory
 from .transfer.emerge.executor import AUTO_CONTENT_TYPE
 from .transfer.emerge.write_intent import WriteIntent
+from .transfer.emerge.unbound_write_intent import UnboundWriteIntentGenerator
 from .transfer.inbound.downloaded_file import DownloadedFile
 from .transfer.outbound.copy_source import CopySource
 from .transfer.outbound.upload_source import UploadSourceBytes, UploadSourceLocalFile, UploadMode
 from .utils import (
     B2TraceMeta,
+    Sha1HexDigest,
     b2_url_encode,
     disable_trace,
     limit_trace_arguments,
@@ -491,6 +493,7 @@ class Bucket(metaclass=B2TraceMeta):
         encryption: Optional[EncryptionSetting] = None,
         file_retention: Optional[FileRetentionSetting] = None,
         legal_hold: Optional[LegalHold] = None,
+        large_file_sha1: Optional[Sha1HexDigest] = None,
     ):
         """
         Upload bytes in memory to a B2 file.
@@ -503,6 +506,7 @@ class Bucket(metaclass=B2TraceMeta):
         :param b2sdk.v2.EncryptionSetting encryption: encryption settings (``None`` if unknown)
         :param b2sdk.v2.FileRetentionSetting file_retention: file retention setting
         :param bool legal_hold: legal hold setting
+        :param Sha1HexDigest,None large_file_sha1: SHA-1 hash of the result file or ``None`` if unknown
         :rtype: generator[b2sdk.v2.FileVersion]
         """
         upload_source = UploadSourceBytes(data_bytes)
@@ -515,6 +519,7 @@ class Bucket(metaclass=B2TraceMeta):
             encryption=encryption,
             file_retention=file_retention,
             legal_hold=legal_hold,
+            large_file_sha1=large_file_sha1,
         )
 
     def upload_local_file(
@@ -581,6 +586,123 @@ class Bucket(metaclass=B2TraceMeta):
             large_file_sha1=large_file_sha1,
         )
 
+    def upload_unbound_stream(
+        self,
+        read_only_object,
+        file_name: str,
+        content_type: str = None,
+        file_info: Optional[Dict[str, str]] = None,
+        progress_listener: Optional[AbstractProgressListener] = None,
+        recommended_upload_part_size: Optional[int] = None,
+        encryption: Optional[EncryptionSetting] = None,
+        file_retention: Optional[FileRetentionSetting] = None,
+        legal_hold: Optional[LegalHold] = None,
+        min_part_size: Optional[int] = None,
+        max_part_size: Optional[int] = None,
+        large_file_sha1: Optional[Sha1HexDigest] = None,
+        buffers_count: int = 2,
+        buffer_size: Optional[int] = None,
+        read_size: int = 8192,
+        unused_buffer_timeout_seconds: float = 3600.0,
+    ):
+        """
+        Upload an unbound file-like read-only object to a B2 file.
+
+        It is assumed that this object is streamed like stdin or socket, and the size is not known up front.
+        It is up to caller to ensure that this object is open and available through the whole streaming process.
+
+        If stdin is to be passed, consider opening it in binary mode, if possible on the platform:
+
+        .. code-block:: python
+
+            with open(sys.stdin.fileno(), mode='rb', buffering=min_part_size, closefd=False) as source:
+                bucket.upload_unbound_stream(source, 'target-file')
+
+        For platforms without file descriptors, one can use the following:
+
+        .. code-block:: python
+
+            bucket.upload_unbound_stream(sys.stdin.buffer, 'target-file')
+
+        but note that buffering in this case depends on the interpreter mode.
+
+        ``min_part_size``, ``recommended_upload_part_size`` and ``max_part_size`` should
+        all be greater than ``account_info.get_absolute_minimum_part_size()``.
+
+        ``buffers_count`` describes a desired number of buffers that are to be used. Minimal amount is two, as we need
+        to determine the method of uploading this stream (if there's only a single buffer we send it as a normal file,
+        if there are at least two – as a large file).
+        Number of buffers determines the amount of memory used by the streaming process and, in turns, describe
+        the amount of data that can be pulled from ``read_only_object`` while also uploading it. Providing multiple
+        buffers also allows for higher parallelization. Default two buffers allow for the process to fill one buffer
+        with data while the other one is being sent to the B2. While only one buffer can be filled with data at once,
+        all others are used to send the data in parallel (limited only by the number of parallel threads).
+        Buffer size can be controlled by ``buffer_size`` parameter. If left unset, it will default to
+        a value of ``recommended_upload_part_size``, whatever it resolves to be.
+        Note that in the current implementation buffers are (almost) directly sent to B2, thus whatever is picked
+        as the ``buffer_size`` will also become the size of the part when uploading a large file in this manner.
+        In rare cases, namely when the whole buffer was sent, but there was an error during sending of last bytes
+        and a retry was issued, another buffer (above the aforementioned limit) will be allocated.
+
+        :param read_only_object: any object containing a ``read`` method accepting size of the read
+        :param file_name: a file name of the new B2 file
+        :param content_type: the MIME type, or ``None`` to accept the default based on file extension of the B2 file name
+        :param file_info: a file info to store with the file or ``None`` to not store anything
+        :param progress_listener: a progress listener object to use, or ``None`` to not report progress
+        :param encryption: encryption settings (``None`` if unknown)
+        :param file_retention: file retention setting
+        :param legal_hold: legal hold setting
+        :param min_part_size: a minimum size of a part
+        :param recommended_upload_part_size: the recommended part size to use for uploading local sources
+                        or ``None`` to determine automatically
+        :param max_part_size: a maximum size of a part
+        :param large_file_sha1: SHA-1 hash of the result file or ``None`` if unknown
+        :param buffers_count: desired number of buffers allocated, cannot be smaller than 2
+        :param buffer_size: size of a single buffer that we pull data to or upload data to B2. If ``None``,
+                        value of ``recommended_upload_part_size`` is used. If that also is ``None``,
+                        it will be determined automatically as "recommended upload size".
+        :param read_size: size of a single read operation performed on the ``read_only_object``
+        :param unused_buffer_timeout_seconds: amount of time that a buffer can be idle before returning error
+        :rtype: b2sdk.v2.FileVersion
+        """
+        if buffers_count <= 1:
+            raise ValueError('buffers_count has to be at least 2')
+        if read_size <= 0:
+            raise ValueError('read_size has to be a positive integer')
+        if unused_buffer_timeout_seconds <= 0.0:
+            raise ValueError('unused_buffer_timeout_seconds has to be a positive float')
+
+        buffer_size = buffer_size or recommended_upload_part_size
+        if buffer_size is None:
+            planner = self.api.services.emerger.get_emerge_planner()
+            buffer_size = planner.recommended_upload_part_size
+
+        return self._create_file(
+            self.api.services.emerger.emerge_unbound,
+            UnboundWriteIntentGenerator(
+                read_only_object,
+                buffer_size,
+                read_size=read_size,
+                queue_size=buffers_count,
+                queue_timeout_seconds=unused_buffer_timeout_seconds,
+            ).iterator(),
+            file_name,
+            content_type=content_type,
+            file_info=file_info,
+            progress_listener=progress_listener,
+            encryption=encryption,
+            file_retention=file_retention,
+            legal_hold=legal_hold,
+            min_part_size=min_part_size,
+            recommended_upload_part_size=recommended_upload_part_size,
+            max_part_size=max_part_size,
+            # This is a parameter for EmergeExecutor.execute_emerge_plan telling
+            # how many buffers in parallel can be handled at once. We ensure that one buffer
+            # is always downloading data from the stream while others are being uploaded.
+            max_queue_size=buffers_count - 1,
+            large_file_sha1=large_file_sha1,
+        )
+
     def upload(
         self,
         upload_source,
@@ -592,6 +714,7 @@ class Bucket(metaclass=B2TraceMeta):
         encryption: Optional[EncryptionSetting] = None,
         file_retention: Optional[FileRetentionSetting] = None,
         legal_hold: Optional[LegalHold] = None,
+        large_file_sha1: Optional[Sha1HexDigest] = None,
     ):
         """
         Upload a file to B2, retrying as needed.
@@ -613,6 +736,7 @@ class Bucket(metaclass=B2TraceMeta):
         :param b2sdk.v2.EncryptionSetting encryption: encryption settings (``None`` if unknown)
         :param b2sdk.v2.FileRetentionSetting file_retention: file retention setting
         :param bool legal_hold: legal hold setting
+        :param Sha1HexDigest,None large_file_sha1: SHA-1 hash of the result file or ``None`` if unknown
         :rtype: b2sdk.v2.FileVersion
         """
         return self.create_file(
@@ -626,6 +750,7 @@ class Bucket(metaclass=B2TraceMeta):
             encryption=encryption,
             file_retention=file_retention,
             legal_hold=legal_hold,
+            large_file_sha1=large_file_sha1,
         )
 
     def create_file(
@@ -766,6 +891,7 @@ class Bucket(metaclass=B2TraceMeta):
         min_part_size=None,
         max_part_size=None,
         large_file_sha1=None,
+        **kwargs
     ):
         validate_b2_file_name(file_name)
         progress_listener = progress_listener or DoNothingProgressListener()
@@ -785,6 +911,7 @@ class Bucket(metaclass=B2TraceMeta):
             min_part_size=min_part_size,
             max_part_size=max_part_size,
             large_file_sha1=large_file_sha1,
+            **kwargs
         )
 
     def concatenate(
@@ -855,7 +982,7 @@ class Bucket(metaclass=B2TraceMeta):
         encryption: Optional[EncryptionSetting] = None,
         file_retention: Optional[FileRetentionSetting] = None,
         legal_hold: Optional[LegalHold] = None,
-        large_file_sha1: Optional[str] = None,
+        large_file_sha1: Optional[Sha1HexDigest] = None,
     ):
         """
         Creates a new file in this bucket by concatenating stream of multiple remote or local sources.