+++ b2sdk/transfer/emerge/exception.py
@@ -1,0 +1,18 @@
+######################################################################
+#
+# File: b2sdk/transfer/emerge/exception.py
+#
+# Copyright 2022 Backblaze Inc. All Rights Reserved.
+#
+# License https://www.backblaze.com/using_b2_code.html
+#
+######################################################################
+
+from b2sdk.exception import B2SimpleError
+
+
+class UnboundStreamBufferTimeout(B2SimpleError):
+    """
+    Raised when there is no space for a new buffer for a certain amount of time.
+    """
+    pass