+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Raising `PotentialS3EndpointPassedAsRealm` when a specific misconfiguration is suspected
 * Add `large_file_sha1` support
 * Add support for incremental upload and sync
+* Ability to stream data from any source into a B2 file
 
 ### Fixed
 * Removed information about replication being in closed beta